refactor: consolidate 3 runCommand clones into one shared util (D3)
gates.ts, dependency-scanner.ts and pipeline.ts each defined a near-identical
runCommand — but pipeline's copy omitted maxBuffer (Node's 1 MB default vs the
others' 5 MB) and the message fallback, a latent truncation hazard despite its
"shared with gates.ts" comment. Extracted src/utils/run-command.ts as the single
source of truth (5 MB default, combined stdout+stderr, message fallback, never
throws) and replaced all three. Call-site timeouts unchanged (all explicit
except gates', which keeps the 60s default). 65/65.
